+++ src/main/java/com/mordenkainen/equivalentenergistics/blocks/condenser/tiles/TileEMCCondenserAdv.java
@@ -20,7 +20,7 @@ public class TileEMCCondenserAdv extends TileEMCCondenser {
     }
 
     @Override
-    protected float getEMCPerTick() {
+    protected double getEMCPerTick() {
         return EqEConfig.emcCondenser.emcPerTick * 10;
     }
